% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFatlasqtl_package.R
\docType{package}
\name{AFatlasqtl-package}
\alias{AFatlasqtl}
\alias{AFatlasqtl-package}
\title{AFatlasqtl: Adaptive-focus implementation of atlasQTL}
\description{
Adaptive-focus strategy of coordinate ascent variational inference (AF-CAVI),a strategy that selectively updates a few instead of all parameters in CAVI, can efficiently boost the scalability and computational efficieny of Bayesian hiearchical models. This package provides an AF-CAVI implementation for atlasQTL, a flexible sparse regression for variable selection in large predictor and response settings, based on a series of hierarchically-related spike-and-slab submodels. This implementation allows for multi-trait joint quantatitive trait loci (QTL) mapping at a scale of thousands of traits and genomic variants, and tens of thousands samples. The atlasQTL model is particularly designed for QTL hotspot (QTL associated with multiple traits) detections with a hiearchial structure.
}
\author{
\strong{Maintainer}: Yiran Li \email{yiran.li@mrc-bsu.cam.ac.uk}

Authors:
\itemize{
  \item Helene Ruffieux \email{helene.ruffieux@mrc-bsu.cam.ac.uk}
}

}
\keyword{internal}
